regenerate rdef to fix toolbar image data
The old BGhostview_x86.rdef file had image data that the internal BBitmap
translator couldn't handle causing BGhostview to give an error about a missing
XPMTranslator.

This commit creates a new generate_rdef.sh script that was used to regenerate
the rdef file.  XPMTranslator must be installed when regenerating the rdef, but
not for building or running the app.

In addition, a few obsolete files were removed and a few were renamed to
clean things up.