Change handling of painter_sem
The "presize" function can be called twice during the rendering of a
page (not sure why, but this is what I see after some debugging). This
is a problem if it tries to acquire the painter_sem when no one has
released it (because display_page was never called).

Instead, create the semaphore with a 0 value and release it when the
page is ready. This is a little simpler, but the design with all these
semaphores everywhere is still not great.